% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2.R
\name{strsplit2}
\alias{strsplit2}
\title{Split a string}
\usage{
strsplit2(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit2(x, split = ",")
}
